backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Svennson17 <kiryl.shvedau@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
psycopg2 = "^2.9.5"
flake8 = "^6.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
qrcode = "^7.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
